Fix bootpd check on macOS >= 15
On macOS >= 15 bootpd is likely allowed to receive incoming connections
as built-in software, and it will not be listed in the allowed
applications. In this case we decide that bootpd is blocked and force
the user to try to add and unblock it, which will never succeed.

Fixed using the new --getallowedsigned option. If the option is enabled,
we know that bootpd is not blocked. If the option is not enabled, or the
fails, we fallback to checking the list.

Tested on macOS 15.3.